+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v52.x-v53 (In Progress)
+- Ancient Cave: Zoo room NPCs removed to prevent hardlocks.
 - Challenges: New challenge mode "Ancient Cave: Short Skips" added.
 - Credits: EBSS AC 2023 winner temple added.
 - Game improvement patches: Ludicrous speed patch now on by default.